Calls to hasPendingMessage() can return false without reading body
If the latest message header read has no body then
HttpMessageStreamReader.hasPendingMessage() now returns "false"
without requiring an attempt to read it by calling readMessage() again.

This behavior is required when no further data is sent upstream
(= no calls to readMessage()) after a GET header and a connection
analyzer uses hasPendingMessage() to check if the request is terminated
in order to start waiting for a response.